dom/media/tests/mochitest/test_peerConnection_bug835370.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=825703
-->
<head>
  <meta charset="utf-8">
    <title>Bug 835370: PeerConnection.createOffer valid/invalid constraints permutations</title>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="head.js"></script>
  </meta>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=825703">PeerConnection.createOffer constraints valid/invalid constraints permutations</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="application/javascript">
if (navigator.platform.startsWith("Win")) {
  SimpleTest.expectAssertions(0, 9);
}
runTest(function () {
  var pconnect  = new mozRTCPeerConnection();
  var pconnects = new mozRTCPeerConnection();

  function step1(offer) {}
  function failed(code) {}

  var exception = null;
  try { pconnects.createOffer(step1, failed); } catch (e) { exception = e; }
  ok(!exception, "createOffer(step1, failed) succeeds");
  exception = null;
  try { pconnect.createOffer(step1, failed, 1); } catch (e) { exception = e; }
  ok(exception, "createOffer(step1, failed, 1) throws");
  exception = null;
  try { pconnect.createOffer(step1, failed, []); } catch (e) { exception = e; }
  ok(exception, "createOffer(step1, failed, []) throws");
  exception = null;
  try { pconnects.createOffer(step1, failed, {}); } catch (e) { exception = e; }
  ok(!exception, "createOffer(step1, failed, {}) succeeds");
  exception = null;
  try { pconnect.createOffer(step1, failed, { mandatory: [] }); } catch (e) { exception = e; }
  ok(exception, "createOffer(step1, failed, { mandatory: [] }) throws");
  exception = null;
  try { pconnect.createOffer(step1, failed, { mandatory: { FooBar: true } }); } catch (e) { exception = e; }
  ok(exception, "createOffer(step1, failed, { mandatory: { FooBar: true } }) throws");
  exception = null;
  try { pconnect.createOffer(step1, failed, { optional: {} }); } catch (e) { exception = e; }
  ok(exception, "createOffer(step1, failed, { optional: {} }) throws");
  exception = null;
  try { pconnects.createOffer(step1, failed, { optional: [] }); } catch (e) { exception = e; }
  ok(!exception, "createOffer(step1, failed, { optional: [] }) succeeds");
  exception = null;
  try { pconnect.createOffer(step1, failed, { optional: [1] }); } catch (e) { exception = e; }
  ok(exception, "createOffer(step1, failed, { optional: [1] }) throws");
  exception = null;
  try { pconnect.createOffer(step1, failed, { optional: [{ OfferToReceiveVideo: false, OfferToReceiveAudio: true, }] }); } catch (e) { exception = e; }
  ok(exception, "createOffer(step1, failed, { optional: [{ OfferToReceiveVideo: false, OfferToReceiveAudio: true, }] }) throws");
  exception = null;
  try { pconnects.createOffer(step1, failed, { mandatory: { OfferToReceiveVideo: false, OfferToReceiveAudio: true, MozDontOfferDataChannel: true}, optional: [{ VoiceActivityDetection: true }, { FooBar: "42"  }] }); } catch (e) { exception = e; }
  ok(!exception, "createOffer(step1, failed, { mandatory: { OfferToReceiveVideo: false, OfferToReceiveAudio: true, MozDontOfferDataChannel: true}, optional: [{ VoiceActivityDetection: true }, { FooBar: \"42\"  }] }) succeeds");
  pconnect = null;
  pconnects = null;
  SimpleTest.finish();
}, true);
</script>
</pre>
</body>
</html>
